Minor cleanup (#715)
* server: refactor bidRespKey

* server: split http handling and functionality
